+++ api/src/main/java/org/openmrs/module/fhir2/api/impl/FhirTaskServiceImpl.java
@@ -74,8 +74,7 @@ public class FhirTaskServiceImpl extends BaseFhirService<Task, FhirTask> impleme
 		        .addParameter(FhirConstants.STATUS_SEARCH_HANDLER, status)
 		        .addParameter(FhirConstants.COMMON_SEARCH_HANDLER, FhirConstants.ID_PROPERTY, id)
 		        .addParameter(FhirConstants.COMMON_SEARCH_HANDLER, FhirConstants.LAST_UPDATED_PROPERTY, lastUpdated)
-			.addParameter(FhirConstants.INCLUDE_SEARCH_HANDLER, includes)
-		        .setSortSpec(sort);
+		        .addParameter(FhirConstants.INCLUDE_SEARCH_HANDLER, includes).setSortSpec(sort);
 		
 		return searchQuery.getQueryResults(theParams, dao, translator, searchQueryInclude);
 	}
