+++ fullstack-integrationtest/src/validation/ CustomClientHttpRequestInterceptor.java
@@ -1,0 +1,34 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.fhir2.fullstack-integrationtest.validation;
+
+import org.springframework.http.client.ClientHttpRequestExecution;
+import org.springframework.http.client.ClientHttpRequestInterceptor;
+import org.springframework.http.client.ClientHttpResponse;
+
+public class  CustomClientHttpRequestInterceptor implements ClientHttpRequestInterceptor {
+    private static Logger LOGGER = LoggerFactory.getLogger(CustomClientHttpRequestInterceptor.class);
+
+    @Override
+    public ClientHttpResponse intercept(
+            HttpRequest request, byte[] body,
+            ClientHttpRequestExecution execution) throws IOException {
+
+    //logRequestDetails(request);
+        return execution.execute(request, body);
+    }
+
+    private void logRequestDetails(HttpRequest request) {
+        LOGGER.info("Headers: {}", request.getHeaders());
+        LOGGER.info("Request Method: {}", request.getMethod());
+        LOGGER.info("Request URI: {}", request.getURI());
+    }
+    
+}
