+++ pom.xml
@@ -44,6 +44,7 @@
         <module>integration-tests</module>
         <module>integration-tests-2.1</module>
         <module>integration-tests-2.2</module>
+        <module>fullstack-integrationtest</module>
     </modules>
 
     <repositories>
